8332900: RISC-V: refactor nativeInst_riscv.cpp and macroAssembler_riscv.cpp
